app/requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
